Revert "Organización de los métodos de resumen"
