+++ dato/models.py
@@ -13,8 +13,8 @@ class Datos(models.Model):
     Temperatura = models.DecimalField(max_digits=6, decimal_places=2, default=0)
     Calor = models.DecimalField(max_digits=6, decimal_places=2, default=0)
     Concentracion = models.DecimalField(max_digits=6, decimal_places=2, default=0)
-    Latitud = models.DecimalField(max_digits=12, decimal_places=7, default=0)
-    Longitud = models.DecimalField(max_digits=12, decimal_places=7, default=0)
+    Latitud = models.DecimalField(max_digits=10, decimal_places=7, default=0)
+    Longitud = models.DecimalField(max_digits=10, decimal_places=7, default=0)
     SensorHumo = models.BooleanField(default=False)
     SensorMetano = models.BooleanField(default=False)
     fecha = models.DateTimeField('Fecha y hora')
